+++ src/org/durka/hallmonitor/DefaultActivity.java
@@ -6,6 +6,7 @@ import java.util.Timer;
 import java.util.TimerTask;
 
 import org.durka.hallmonitor.Functions.Actions;
+import org.durka.hallmonitor.Functions.TorchActions;
 
 import android.app.Activity;
 import android.appwidget.AppWidgetHostView;
@@ -64,6 +65,7 @@ public class DefaultActivity extends Activity {
     
     //all the views we need
     public ImageButton torchButton = null;
+    public ImageButton torchButton2 = null;
     private ImageButton cameraButton = null;
 
 	//we need to kill this activity when the screen opens
@@ -210,7 +212,15 @@ public class DefaultActivity extends Activity {
 			torchButton.setVisibility(View.INVISIBLE);
 		}
 		
-		//hide or show the torch button as required
+		//hide or show the alternate torch button as required
+		if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("pref_flash_controls_alternative", false))
+		{
+			torchButton2.setVisibility(View.VISIBLE);
+		} else {
+			torchButton2.setVisibility(View.INVISIBLE);
+		}
+		
+		//hide or show the camera button as required
 		if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("pref_camera_controls", false))
 		{
 			cameraButton.setVisibility(View.VISIBLE);
@@ -330,8 +340,17 @@ public class DefaultActivity extends Activity {
 		Functions.Actions.toggle_torch(this);
 	}
 	
+	//toggle the alternative torch
+	public void toggleTorch(View view) {
+		Functions.Actions.toggle_torch_alternative(this);
+	}
+	
 	//fire up the camera
 	public void camera_start(View view) {
+		if (Functions.flashIsOn) {
+		 	TorchActions.turnOffFlash();
+		 	torchButton2.setImageResource(R.drawable.ic_appwidget_torch_off);
+		}
 		Functions.Actions.start_camera(this);
 	}
 	
@@ -387,6 +406,7 @@ public class DefaultActivity extends Activity {
 		
 		//get the views we need
 	    torchButton = (ImageButton) findViewById(R.id.torchbutton);
+	    torchButton2 = (ImageButton) findViewById(R.id.torchbutton2);
 	    cameraButton = (ImageButton) findViewById(R.id.camerabutton);
 	}
 
@@ -454,10 +474,15 @@ public class DefaultActivity extends Activity {
 	protected void onStop() {
 		super.onStop();
 		Log.d("DA-oS", "stopping");
+		if (PreferenceManager.getDefaultSharedPreferences(this).getBoolean("pref_keyguard", true)) {
+			this.getWindow().addFlags(WindowManager.LayoutParams.FLAG_DISMISS_KEYGUARD);
+		}
 		if (Functions.Actions.timerTask != null) {
 			Functions.Actions.timerTask.cancel();
 		}
-
+		if (Functions.flashIsOn) {
+				TorchActions.turnOffFlash();
+		}
 		if (camera_up) {
 			Functions.Actions.end_camera(this, false);
 		}